requirements.txt
tensorflow>=2.16.1
opencv-python>=4.9.0
pandas>=2.1.4
scikit-learn>=1.3.2
pytest>=8.0.0
Pillow>=10.2.0
matplotlib>=3.8.2
streamlit>=1.32.0
seaborn>=0.13.0